util.py split into citekey.py and csl_dict.py
