+++ .gitignore
@@ -4,6 +4,7 @@ __pycache__/
 .ipynb_checkpoints
 .cache
 .pytest_cache/
+.idea/
 
 # System specific files
