+++ manubot/cite/citekey.py
@@ -1,38 +1,7 @@
 import functools
 import logging
-import re
 
-from manubot.util import import_function
-
-citeproc_retrievers = {
-    'doi': 'manubot.cite.doi.get_doi_csl_item',
-    'pmid': 'manubot.cite.pubmed.get_pubmed_csl_item',
-    'pmcid': 'manubot.cite.pubmed.get_pmc_csl_item',
-    'arxiv': 'manubot.cite.arxiv.get_arxiv_csl_item',
-    'isbn': 'manubot.cite.isbn.get_isbn_csl_item',
-    'wikidata': 'manubot.cite.wikidata.get_wikidata_csl_item',
-    'url': 'manubot.cite.url.get_url_csl_item',
-}
-
-"""
-Regex to extract citation keys.
-The leading '@' is omitted from the single match group.
-
-Same rules as pandoc, except more permissive in the following ways:
-
-1. the final character can be a slash because many URLs end in a slash.
-2. underscores are allowed in internal characters because URLs, DOIs, and
-   citation tags often contain underscores.
-
-If a citekey does not match this regex, it can be substituted for a
-tag that does, as defined in citation-tags.tsv.
-
-https://github.com/greenelab/manubot-rootstock/issues/2#issuecomment-312153192
-
-Prototyped at https://regex101.com/r/s3Asz3/4
-"""
-citekey_pattern = re.compile(
-    r'(?<!\w)@([a-zA-Z0-9][\w:.#$%&\-+?<>~/]*[a-zA-Z0-9/])')
+#from manubot.cite.util import shorten_citekey, citeproc_retrievers
 
 
 @functools.lru_cache(maxsize=5_000)
@@ -71,6 +40,68 @@ def standardize_citekey(citekey, warn_if_changed=False):
     return standard_citekey
 
 
+# NOTE: tentatively - must split to individual functions correspnding to 
+#       methods of CSL_dict class 
+def citekey_to_csl_item(citekey, prune=True):
+    """
+    Generate a CSL Item (Python dictionary) for the input citekey.
+    """
+    citekey == standardize_citekey(citekey, warn_if_changed=True)
+    source, identifier = citekey.split(':', 1)
+
+    if source in citeproc_retrievers:
+        citeproc_retriever = import_function(citeproc_retrievers[source])
+        csl_item = citeproc_retriever(identifier)
+    else:
+        msg = f'Unsupported citation source {source!r} in {citekey!r}'
+        raise ValueError(msg)
+
+    from manubot import __version__ as manubot_version
+    from manubot.cite.citeproc import (
+        csl_item_passthrough,
+        append_to_csl_item_note,
+    )
+
+    note_text = f'This CSL JSON Item was automatically generated by Manubot v{manubot_version} using citation-by-identifier.'
+    note_dict = {
+        'standard_id': citekey,
+    }
+    append_to_csl_item_note(csl_item, note_text, note_dict)
+
+    short_citekey = shorten_citekey(citekey)
+    csl_item = csl_item_passthrough(csl_item, set_id=short_citekey, prune=prune)
+
+    return csl_item
+
+import functools
+import logging
+import re
+
+# FIXME: this goes away with new classes
+from manubot.util import import_function
+
+
+"""
+Regex to extract citation keys.
+The leading '@' is omitted from the single match group.
+
+Same rules as pandoc, except more permissive in the following ways:
+
+1. the final character can be a slash because many URLs end in a slash.
+2. underscores are allowed in internal characters because URLs, DOIs, and
+   citation tags often contain underscores.
+
+If a citekey does not match this regex, it can be substituted for a
+tag that does, as defined in citation-tags.tsv.
+
+https://github.com/greenelab/manubot-rootstock/issues/2#issuecomment-312153192
+
+Prototyped at https://regex101.com/r/s3Asz3/4
+"""
+citekey_pattern = re.compile(
+    r'(?<!\w)@([a-zA-Z0-9][\w:.#$%&\-+?<>~/]*[a-zA-Z0-9/])')
+
+
 regexes = {
     'pmid': re.compile(r'[1-9][0-9]{0,7}'),
     'pmcid': re.compile(r'PMC[0-9]+'),
@@ -80,6 +111,20 @@ regexes = {
 }
 
 
+# FIXME: this goes away with new classes
+citeproc_retrievers = {
+    'doi': 'manubot.cite.doi.get_doi_csl_item',
+    'pmid': 'manubot.cite.pubmed.get_pubmed_csl_item',
+    'pmcid': 'manubot.cite.pubmed.get_pmc_csl_item',
+    'arxiv': 'manubot.cite.arxiv.get_arxiv_csl_item',
+    'isbn': 'manubot.cite.isbn.get_isbn_csl_item',
+    'wikidata': 'manubot.cite.wikidata.get_wikidata_csl_item',
+    'url': 'manubot.cite.url.get_url_csl_item',
+}
+
+
+# FIXME: inspect fucntionality goes to individual Handle 
+#        classes, not if's will be needed
 def inspect_citekey(citekey):
     """
     Check citekeys adhere to expected formats. If an issue is detected a
@@ -256,38 +301,6 @@ def shorten_citekey(standard_citekey):
     return short_citekey
 
 
-def citekey_to_csl_item(citekey, prune=True):
-    """
-    Generate a CSL Item (Python dictionary) for the input citekey.
-    """
-    citekey == standardize_citekey(citekey, warn_if_changed=True)
-    source, identifier = citekey.split(':', 1)
-
-    if source in citeproc_retrievers:
-        citeproc_retriever = import_function(citeproc_retrievers[source])
-        csl_item = citeproc_retriever(identifier)
-    else:
-        msg = f'Unsupported citation source {source!r} in {citekey!r}'
-        raise ValueError(msg)
-
-    from manubot import __version__ as manubot_version
-    from manubot.cite.citeproc import (
-        csl_item_passthrough,
-        append_to_csl_item_note,
-    )
-
-    note_text = f'This CSL JSON Item was automatically generated by Manubot v{manubot_version} using citation-by-identifier.'
-    note_dict = {
-        'standard_id': citekey,
-    }
-    append_to_csl_item_note(csl_item, note_text, note_dict)
-
-    short_citekey = shorten_citekey(citekey)
-    csl_item = csl_item_passthrough(csl_item, set_id=short_citekey, prune=prune)
-
-    return csl_item
-
-
 def infer_citekey_prefix(citekey):
     """
     Passthrough citekey if it has a valid citation key prefix. Otherwise,
@@ -300,59 +313,4 @@ def infer_citekey_prefix(citekey):
             return citekey
         if citekey.lower().startswith(prefix):
             return prefix + citekey[len(prefix):]
-    return f'raw:{citekey}'
-
-
-def csl_item_set_standard_id(csl_item):
-    """
-    Extract the standard_id (standard citation key) for a csl_item and modify the csl_item in-place to set its "id" field.
-    The standard_id is extracted from a "standard_citation" field, the "note" field, or the "id" field.
-    If extracting the citation from the "id" field, uses the infer_citekey_prefix function to set the prefix.
-    For example, if the extracted standard_id does not begin with a supported prefix (e.g. "doi:", "pmid:"
-    or "raw:"), the citation is assumed to be raw and given a "raw:" prefix. The extracted citation
-    (referred to as "original_standard_id") is checked for validity and standardized, after which it is
-    the final "standard_id".
-
-    Regarding csl_item modification, the csl_item "id" field is set to the standard_citation and the note field
-    is created or updated with key-value pairs for standard_id, original_standard_id, and original_id.
-
-    Note that the Manubot software generally refers to the "id" of a CSL Item as a citekey.
-    However, in this context, we use "id" rather than "citekey" for consistency with CSL's "id" field.
-    """
-    if not isinstance(csl_item, dict):
-        raise ValueError("csl_item must be a CSL Data Item represented as a Python dictionary")
-
-    from manubot.cite.citeproc import (
-        append_to_csl_item_note,
-        parse_csl_item_note,
-    )
-    note_dict = parse_csl_item_note(csl_item.get('note', ''))
-
-    original_id = None
-    original_standard_id = None
-    if 'id' in csl_item:
-        original_id = csl_item['id']
-        original_standard_id = infer_citekey_prefix(original_id)
-    if 'standard_id' in note_dict:
-        original_standard_id = note_dict['standard_id']
-    if 'standard_citation' in csl_item:
-        original_standard_id = csl_item.pop('standard_citation')
-    if original_standard_id is None:
-        raise ValueError(
-            'csl_item_set_standard_id could not detect a field with a citation / standard_citation. '
-            'Consider setting the CSL Item "id" field.'
-        )
-    assert is_valid_citekey(original_standard_id, allow_raw=True)
-    standard_id = standardize_citekey(original_standard_id, warn_if_changed=False)
-    add_to_note = {}
-    if original_id and original_id != standard_id:
-        if original_id != note_dict.get('original_id'):
-            add_to_note['original_id'] = original_id
-    if original_standard_id and original_standard_id != standard_id:
-        if original_standard_id != note_dict.get('original_standard_id'):
-            add_to_note['original_standard_id'] = original_standard_id
-    if standard_id != note_dict.get('standard_id'):
-        add_to_note['standard_id'] = standard_id
-    append_to_csl_item_note(csl_item, dictionary=add_to_note)
-    csl_item['id'] = standard_id
-    return csl_item
+    return f'raw:{citekey}'
